+++ plugin/src/main/kotlin/org/example/githubpang/rossynt/services/RestartableBackendService.kt
@@ -29,7 +29,7 @@ internal class RestartableBackendService : IBackendService {
         destroyBackendService()
     }
 
-    override suspend fun compileFile(fileText: String?, filePath: String?): TreeNode? = backendService?.compileFile(fileText, filePath)
+    override suspend fun compileFile(fileText: String?, filePath: String?, cSharpVersion: CSharpVersion): TreeNode? = backendService?.compileFile(fileText, filePath, cSharpVersion)
     override suspend fun getNodeInfo(nodeId: String): Map<String, String> = backendService?.getNodeInfo(nodeId) ?: HashMap()
     override suspend fun findNode(start: Int, end: Int): String? = backendService?.findNode(start, end)